requirements.txt
ccxt==4.1.13
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
websocket-client==1.7.0
tensorflow==2.15.0
scikit-learn==1.3.2
pytest==7.4.3
python-binance==1.0.19
aiohttp==3.9.1
fastapi==0.105.0
uvicorn==0.24.0
python-telegram-bot==20.7